Changed views that use GET requests to delete objects to use POST.
Fixes #1796.
 - Legacy-Id: 13654
Note: SVN reference [13615] has been migrated to Git commit 3ccb550888ce72355c67c209905cf768e01724fd

Note: SVN reference [13627] has been migrated to Git commit b908caf95c0f1a7a28498222f2019511c45f600f

+++ ietf/secr/groups/views.py
@@ -190,14 +190,17 @@ def delete_role(request, acronym, id):
     """
     group = get_object_or_404(Group, acronym=acronym)
     role = get_object_or_404(Role, id=id)
+    
+    if request.method == 'POST' and request.POST['post'] == 'yes':
+        # save group
+        save_group_in_history(group)
 
-    # save group
-    save_group_in_history(group)
+        role.delete()
+        messages.success(request, 'The entry was deleted successfully')
+        return redirect('ietf.secr.groups.views.people', acronym=acronym)
 
-    role.delete()
+    return render(request, 'confirm_delete.html', {'object': role})
 
-    messages.success(request, 'The entry was deleted successfully')
-    return redirect('ietf.secr.groups.views.people', acronym=acronym)
 
 @role_required('Secretariat')
 def edit(request, acronym):
